+++ annotations/src/main/java/com/karuslabs/annotations/Immutable.java
@@ -27,7 +27,6 @@ import java.lang.annotation.*;
 
 import static java.lang.annotation.RetentionPolicy.SOURCE;
 
-
 @Documented
 @Retention(SOURCE)
 public @interface Immutable {