+++ annotations/src/main/java/com/karuslabs/annotations/Static.java
@@ -28,7 +28,6 @@ import java.lang.annotation.*;
 import static java.lang.annotation.ElementType.TYPE;
 import static java.lang.annotation.RetentionPolicy.SOURCE;
 
-
 @Documented
 @Retention(SOURCE)
 @Target(TYPE)