+++ annotations/src/test/java/com/karuslabs/annotations/processor/MessagesTest.java
@@ -27,7 +27,6 @@ import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-
 class MessagesTest {
     
     @Test