+++ commons/src/main/java/com/karuslabs/commons/command/aot/lexers/CommandLexer.java
@@ -23,7 +23,8 @@
  */
 package com.karuslabs.commons.command.aot.lexers;
 
-import com.karuslabs.commons.command.aot.*;
+import com.karuslabs.annotations.processor.Logger;
+import com.karuslabs.commons.command.aot.Token;
 
 import java.util.*;