+++ commons/src/main/java/com/karuslabs/commons/command/aot/lexers/Lexer.java
@@ -23,6 +23,7 @@
  */
 package com.karuslabs.commons.command.aot.lexers;
 
+import com.karuslabs.annotations.processor.Logger;
 import com.karuslabs.commons.command.aot.*;
 import com.karuslabs.commons.command.aot.Identifier.Type;